feat: Add sort & filter bottom sheet for doctors
- Implement specialization & rating filters with API integration
- Refactor bottom sheet into reusable components
- Separate state management for Home and SeeAll screens
- Support filter persistence with pagination

+++ lib/features/home/data/models/recommendation_doctors_response.dart
@@ -54,7 +54,7 @@ class DoctorData {
   final String fullName;
   final String specializationName;
   final String address;
-  final int rate;
+  final double rate;
   final int reviewsCount;
 
   DoctorData({
